inspector: handle unknown scripts
If the debugger pauses on an unknown script it will automatically step in
until a known script is found (or the method returns back). This handles
scenarios with evaluated scripts:

```
function bar() {
    return 4;
}

eval(`
function baz() {
  console.log('baz');
}

function foo() {
  baz();
  baz();
  bar();
  return 5;
}`);
foo();
```

Where stepping into `foo` will keep stepping in until the first known script,
in this case the `console` property getter. Each time the debugger pauses
inside of the unknown script it will immediately try to step into the next
function.

In cases where the user is attempting to step out, the debugger will continue
stepping out until the next known script is encountered. This allows the user
to step into the evaluated code, realize that they didn't want to be there and
immediately step out of the entire frame.